ai-service/requirements.txt
flask
flask-cors
transformers
torch
pydub
speechrecognition
requests
gunicorn
accelerate
python-dotenv